crda: Trust regulatory.bin from Seth Forshee
I have taken over maintenance of the regulatory database from
linville. Add my key so that regulatory.bin files signed by me
will be trusted, but also keep John's key so that existing files
will continue to be trusted. Update the makefile to point to the
new wireless-regdb repository location, and update the readme and
man page to reflect the change in maintainership.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>
Signed-off-by: Luis R. Rodriguez <mcgrof@do-not-panic.com>